enh - revamp website

+++ doc/convert_notebooks.py
@@ -11,7 +11,7 @@ EXAMPLES_DIR = os.path.abspath(
     os.path.join(os.path.dirname(__file__), "..", "examples")
 )
 TUTORIALS_DIR = os.path.abspath(
-    os.path.join(os.path.dirname(__file__), "pages", "tutorials")
+    os.path.join(os.path.dirname(__file__), "tutorials")
 )
 
 os.makedirs(TUTORIALS_DIR, exist_ok=True)
@@ -19,8 +19,12 @@ os.makedirs(TUTORIALS_DIR, exist_ok=True)
 
 def clean_tutorials_dir() -> None:
     """Remove previously generated RST files and support png's."""
-    # Delete every generated .rst file in the tutorials output dir
+    # Delete every generated .rst file in the tutorials output dir,
+    # but not the index.rst.
     for f in glob.glob(os.path.join(TUTORIALS_DIR, "*.rst")):
+        # Skip the index.rst
+        if os.path.basename(f) == "index.rst":
+            continue
         os.remove(f)
     # Delete every notebook-support directory that ends with "_files/".
     # This deletes all the png's
